# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FairLaunchCoin {
  /** Fair-launch rank by market cap */
  rank: number;
  /** CoinGecko coin ID */
  id: string;
  name: string;
  symbol: string;
  price?: number | null;
  marketCap?: number | null;
  circulatingSupply?: number | null;
  change24h?: number | null;
  launchYear: number;
  /** e.g. PoW, PoS */
  consensusType: string;
  /** Mining or consensus algorithm name */
  algorithm: string;
  /** Short explanation of fair launch credential */
  whyFair: string;
  imageUrl?: string | null;
  /** Highlighted coin (e.g. Crypton CRP) */
  isFeatured: boolean;
  website?: string | null;
  explorer?: string | null;
  github?: string | null;
  /** Annual staking/yield APY percentage, null if not applicable */
  stakingApy?: number | null;
  /** Type of yield mechanism (e.g. PoS Staking, Node Operator, Forging) */
  yieldType?: string | null;
  /** Latest software release tag (e.g. v1.3.1278) */
  softwareVersion?: string | null;
  /** ISO date string of the latest software release */
  lastReleasedAt?: string | null;
  /** Active node/miner count (available for CRP via Utopia explorer) */
  activeNodes?: number | null;
}

export interface FairLaunchCoinsResponse {
  coins: FairLaunchCoin[];
  /** ISO timestamp of last data fetch */
  lastUpdated: string;
  totalCoins: number;
}

export interface CoinChartResponse {
  id: string;
  /** Array of [timestamp_ms, price_usd] tuples */
  prices: number[][];
  hasData: boolean;
}

export interface FairLaunchStats {
  totalMarketCap: number;
  totalCoins: number;
  topCoin: string;
  /** Current rank of Crypton (CRP) in the fair-launch list */
  cryptonRank?: number | null;
  lastUpdated: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetFairLaunchCoinsParams = {
  /**
   * Optional search query to filter coins by name or symbol
   */
  search?: string;
};