# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: fairlaunch
    description: Fair launch cryptocurrency operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /fairlaunch/coins:
    get:
      operationId: getFairLaunchCoins
      tags: [fairlaunch]
      summary: Get fair launch coins
      description: Returns ranked list of fair-launch cryptocurrencies with live market data from CoinGecko
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Optional search query to filter coins by name or symbol
      responses:
        "200":
          description: List of fair launch coins ranked by market cap
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FairLaunchCoinsResponse"
        "500":
          description: Error fetching data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /fairlaunch/coins/{id}/chart:
    get:
      operationId: getCoinChart
      tags: [fairlaunch]
      summary: Get 7-day price chart for a coin
      description: Returns hourly price data points for the last 7 days. Returns empty array for coins without chart data.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Coin ID (matches the id field from the coins list)
      responses:
        "200":
          description: Price chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinChartResponse"
        "404":
          description: Coin not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /fairlaunch/stats:
    get:
      operationId: getFairLaunchStats
      tags: [fairlaunch]
      summary: Get summary stats
      description: Returns aggregate statistics for all tracked fair-launch coins
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FairLaunchStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    FairLaunchCoin:
      type: object
      properties:
        rank:
          type: integer
          description: Fair-launch rank by market cap
        id:
          type: string
          description: CoinGecko coin ID
        name:
          type: string
        symbol:
          type: string
        price:
          type: number
          nullable: true
        marketCap:
          type: number
          nullable: true
        circulatingSupply:
          type: number
          nullable: true
        change24h:
          type: number
          nullable: true
        launchYear:
          type: integer
        consensusType:
          type: string
          description: e.g. PoW, PoS
        algorithm:
          type: string
          description: Mining or consensus algorithm name
        whyFair:
          type: string
          description: Short explanation of fair launch credential
        imageUrl:
          type: string
          nullable: true
        isFeatured:
          type: boolean
          description: Highlighted coin (e.g. Crypton CRP)
        website:
          type: string
          nullable: true
        explorer:
          type: string
          nullable: true
        github:
          type: string
          nullable: true
        stakingApy:
          type: number
          nullable: true
          description: Annual staking/yield APY percentage, null if not applicable
        yieldType:
          type: string
          nullable: true
          description: Type of yield mechanism (e.g. PoS Staking, Node Operator, Forging)
        softwareVersion:
          type: string
          nullable: true
          description: Latest software release tag (e.g. v1.3.1278)
        lastReleasedAt:
          type: string
          nullable: true
          description: ISO date string of the latest software release
        activeNodes:
          type: number
          nullable: true
          description: Active node/miner count (available for CRP via Utopia explorer)
      required:
        - rank
        - id
        - name
        - symbol
        - launchYear
        - consensusType
        - algorithm
        - whyFair
        - isFeatured

    FairLaunchCoinsResponse:
      type: object
      properties:
        coins:
          type: array
          items:
            $ref: "#/components/schemas/FairLaunchCoin"
        lastUpdated:
          type: string
          description: ISO timestamp of last data fetch
        totalCoins:
          type: integer
      required:
        - coins
        - lastUpdated
        - totalCoins

    CoinChartResponse:
      type: object
      properties:
        id:
          type: string
        prices:
          type: array
          description: Array of [timestamp_ms, price_usd] tuples
          items:
            type: array
            items:
              type: number
        hasData:
          type: boolean
      required:
        - id
        - prices
        - hasData

    FairLaunchStats:
      type: object
      properties:
        totalMarketCap:
          type: number
        totalCoins:
          type: integer
        topCoin:
          type: string
        cryptonRank:
          type: integer
          nullable: true
          description: Current rank of Crypton (CRP) in the fair-launch list
        lastUpdated:
          type: string
      required:
        - totalMarketCap
        - totalCoins
        - topCoin
        - lastUpdated

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message