# lib/api-zod/src/generated/types/fairLaunchCoin.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FairLaunchCoin {
  /** Fair-launch rank by market cap */
  rank: number;
  /** CoinGecko coin ID */
  id: string;
  name: string;
  symbol: string;
  price?: number | null;
  marketCap?: number | null;
  circulatingSupply?: number | null;
  change24h?: number | null;
  launchYear: number;
  /** e.g. PoW, PoS */
  consensusType: string;
  /** Mining or consensus algorithm name */
  algorithm: string;
  /** Short explanation of fair launch credential */
  whyFair: string;
  imageUrl?: string | null;
  /** Highlighted coin (e.g. Crypton CRP) */
  isFeatured: boolean;
  website?: string | null;
  explorer?: string | null;
  github?: string | null;
  /** Annual staking/yield APY percentage, null if not applicable */
  stakingApy?: number | null;
  /** Type of yield mechanism (e.g. PoS Staking, Node Operator, Forging) */
  yieldType?: string | null;
  /** Latest software release tag (e.g. v1.3.1278) */
  softwareVersion?: string | null;
  /** ISO date string of the latest software release */
  lastReleasedAt?: string | null;
  /** Active node/miner count (available for CRP via Utopia explorer) */
  activeNodes?: number | null;
}